backend_django/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
python-socketio[asgi]
uvicorn
gunicorn
python-dotenv
asgiref
djangorestframework-simplejwt
google-generativeai
requests
dj-database-url
psycopg2-binary
pymongo[srv]==4.6.1
dnspython==2.4.2
